Refactor RightPanelTest for local GUI testing only
