--- src/test/java/spleetwaise/transaction/ui/GuiUnitTest.java
@@ -1,32 +1,0 @@
-package spleetwaise.transaction.ui;
-
-import java.util.Optional;
-
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-import guitests.GuiRobot;
-import guitests.guihandles.exceptions.NodeNotFoundException;
-import javafx.scene.Node;
-import spleetwaise.address.testutil.UiPartExtension;
-
-/**
- * A GUI unit test class for AddressBook.
- */
-public abstract class GuiUnitTest {
-
-    @RegisterExtension
-    public final UiPartExtension uiPartExtension = new UiPartExtension();
-
-    protected final GuiRobot guiRobot = new GuiRobot();
-
-    /**
-     * Retrieves the {@code query} node owned by the {@code rootNode}.
-     *
-     * @param query name of the CSS selector of the node to retrieve.
-     * @throws NodeNotFoundException if no such node exists.
-     */
-    protected <T extends Node> T getChildNode(Node rootNode, String query) {
-        Optional<T> node = guiRobot.from(rootNode).lookup(query).tryQuery();
-        return node.orElseThrow(NodeNotFoundException::new);
-    }
-}
